move adapters to namespace
