+++ lib/coverband.rb
@@ -4,8 +4,8 @@ require 'aws-sdk'
 
 require 'coverband/version'
 require 'coverband/configuration'
-require 'coverband/redis_store'
-require 'coverband/memory_cache_store'
+require 'coverband/adapters/redis_store'
+require 'coverband/adapters/memory_cache_store'
 require 'coverband/base'
 require 'coverband/baseline'
 require 'coverband/reporter'
