+++ test/unit/base_test.rb
@@ -5,13 +5,13 @@ class BaseTest < Test::Unit::TestCase
 
   test 'defaults to a redis store' do
     coverband = Coverband::Base.instance.reset_instance
-    assert_equal Coverband::RedisStore, coverband.instance_variable_get('@reporter').class
+    assert_equal Coverband::Adapters::RedisStore, coverband.instance_variable_get('@reporter').class
   end
 
   test 'configure memory caching' do
     Coverband.configuration.memory_caching = true
     coverband = Coverband::Base.instance.reset_instance
-    assert_equal Coverband::MemoryCacheStore, coverband.instance_variable_get('@reporter').class
+    assert_equal Coverband::Adapters::MemoryCacheStore, coverband.instance_variable_get('@reporter').class
     Coverband.configuration.memory_caching = false
   end
 
@@ -72,7 +72,7 @@ class BaseTest < Test::Unit::TestCase
     coverband = Coverband::Base.instance.reset_instance
     coverband.instance_variable_set("@sample_percentage", 100.0)
     coverband.instance_variable_set("@verbose", true)
-    store = Coverband::RedisStore.new(Redis.new)
+    store = Coverband::Adapters::RedisStore.new(Redis.new)
     coverband.instance_variable_set("@reporter", store)
     store.expects(:store_report).once.with(has_entries(dog_file => { 3 => 5 }) )
     assert_equal false, coverband.instance_variable_get("@enabled")
